feat: 5020 - unselect+edit buttons on OCR page
New files:
* `product_image_button.dart`: Abstract Product Image Button.
* `product_image_crop_button.dart`: Product Image Button editing the current image. Code used to be in `product_image_viewer.dart`. Fixed a bug meanwhile, around `_getCropRect`.
* `product_image_unselect_button.dart`: Product Image Button unselecting the current image. Code used to be in `product_image_viewer.dart`.

Impacted files:
* `edit_ocr_page.dart`: added unselect and edit buttons when relevant; refactored using new enum `ProductImageButtonType`
* `product_image_local_button.dart`: refactored extending new class `ProductImageButton`
* `product_image_server_button.dart`: refactored extending new class `ProductImageButton`
* `product_image_viewer.dart`: moved code to new classes `ProductImageCropButton` and `ProductImageUnselectButton`; refactored using new enum `ProductImageButtonType`
