+++ packages/smooth_app/lib/pages/product/product_image_viewer.dart
@@ -1,28 +1,17 @@
-import 'dart:io';
-
-import 'package:crop_image/crop_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:photo_view/photo_view.dart';
 import 'package:provider/provider.dart';
-import 'package:smooth_app/background/background_task_unselect.dart';
 import 'package:smooth_app/data_models/product_image_data.dart';
 import 'package:smooth_app/data_models/up_to_date_mixin.dart';
-import 'package:smooth_app/database/dao_int.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/database/transient_file.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
-import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
 import 'package:smooth_app/generic_lib/widgets/language_selector.dart';
 import 'package:smooth_app/generic_lib/widgets/picture_not_found.dart';
 import 'package:smooth_app/helpers/product_cards_helper.dart';
-import 'package:smooth_app/pages/crop_page.dart';
-import 'package:smooth_app/pages/image_crop_page.dart';
-import 'package:smooth_app/pages/product/common/product_refresher.dart';
-import 'package:smooth_app/pages/product/edit_image_button.dart';
-import 'package:smooth_app/pages/product/product_image_local_button.dart';
-import 'package:smooth_app/pages/product/product_image_server_button.dart';
+import 'package:smooth_app/pages/product/product_image_button.dart';
 
 /// Displays a full-screen image with an "edit" floating button.
 class ProductImageViewer extends StatefulWidget {
@@ -53,6 +42,16 @@ class _ProductImageViewerState extends State<ProductImageViewer>
     super.initState();
     initUpToDate(widget.product, context.read<LocalDatabase>());
   }
+
+  Widget _getImageButton(final ProductImageButtonType type) => Padding(
+        padding: const EdgeInsets.symmetric(horizontal: SMALL_SPACE),
+        child: type.getButton(
+          product: upToDateProduct,
+          imageField: widget.imageField,
+          language: widget.language,
+          isLoggedInMandatory: widget.isLoggedInMandatory,
+        ),
+      );
 
   @override
   Widget build(BuildContext context) {
@@ -174,26 +173,10 @@ class _ProductImageViewerState extends State<ProductImageViewer>
             mainAxisSize: MainAxisSize.max,
             children: <Widget>[
               Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: SMALL_SPACE),
-                  child: ProductImageServerButton(
-                    product: upToDateProduct,
-                    imageField: widget.imageField,
-                    language: widget.language,
-                    isLoggedInMandatory: widget.isLoggedInMandatory,
-                  ),
-                ),
+                child: _getImageButton(ProductImageButtonType.server),
               ),
               Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: SMALL_SPACE),
-                  child: ProductImageLocalButton(
-                    barcode: barcode,
-                    imageField: widget.imageField,
-                    language: widget.language,
-                    isLoggedInMandatory: widget.isLoggedInMandatory,
-                  ),
-                ),
+                child: _getImageButton(ProductImageButtonType.local),
               ),
             ],
           ),
@@ -204,18 +187,10 @@ class _ProductImageViewerState extends State<ProductImageViewer>
               mainAxisSize: MainAxisSize.max,
               children: <Widget>[
                 Expanded(
-                  child: Padding(
-                    padding:
-                        const EdgeInsets.symmetric(horizontal: SMALL_SPACE),
-                    child: _getUnselectImageButton(appLocalizations),
-                  ),
+                  child: _getImageButton(ProductImageButtonType.unselect),
                 ),
                 Expanded(
-                  child: Padding(
-                    padding:
-                        const EdgeInsets.symmetric(horizontal: SMALL_SPACE),
-                    child: _getEditImageButton(appLocalizations),
-                  ),
+                  child: _getImageButton(ProductImageButtonType.edit),
                 ),
               ],
             ),
@@ -224,199 +199,9 @@ class _ProductImageViewerState extends State<ProductImageViewer>
     );
   }
 
-  // TODO(monsieurtanuki): refactor as ProductImageCropButton
-  Widget _getEditImageButton(final AppLocalizations appLocalizations) =>
-      EditImageButton(
-        iconData: Icons.edit,
-        label: appLocalizations.edit_photo_button_label,
-        onPressed: _actionEditImage,
-      );
-
-  // TODO(monsieurtanuki): refactor as ProductImageUnselectButton
-  Widget _getUnselectImageButton(final AppLocalizations appLocalizations) =>
-      EditImageButton(
-        iconData: Icons.do_disturb_on,
-        label: appLocalizations.edit_photo_unselect_button_label,
-        onPressed: () => _actionUnselect(appLocalizations),
-      );
-
-  Future<File?> _actionEditImage() async {
-    final NavigatorState navigatorState = Navigator.of(context);
-    if (!await ProductRefresher().checkIfLoggedIn(
-      context,
-      isLoggedInMandatory: widget.isLoggedInMandatory,
-    )) {
-      return null;
-    }
-    // best possibility: with the crop parameters
-    // TODO(monsieurtanuki): maybe we should keep the big image locally, in order to avoid the server call?
-    final ProductImage? productImage = _getBestProductImage();
-    if (productImage != null) {
-      final int? imageId = int.tryParse(productImage.imgid!);
-      if (imageId != null) {
-        return _openEditCroppedImage(imageId, productImage);
-      }
-    }
-
-    // alternate option: use the transient file.
-    File? imageFile = _getTransientFile().getImage();
-    if (imageFile != null) {
-      return _openCropPage(navigatorState, imageFile);
-    }
-
-    // but if not possible, get the best picture from the server.
-    final String? imageUrl = _imageData.getImageUrl(ImageSize.ORIGINAL);
-    if (context.mounted) {
-      imageFile = await downloadImageUrl(
-        context,
-        imageUrl,
-        DaoInt(context.read<LocalDatabase>()),
-      );
-    }
-    if (imageFile != null) {
-      return _openCropPage(navigatorState, imageFile);
-    }
-
-    return null;
-  }
-
   TransientFile _getTransientFile() => TransientFile.fromProductImageData(
         _imageData,
         barcode,
         widget.language,
       );
-
-  Future<void> _actionUnselect(final AppLocalizations appLocalizations) async {
-    final NavigatorState navigatorState = Navigator.of(context);
-
-    if (!await ProductRefresher().checkIfLoggedIn(
-      context,
-      isLoggedInMandatory: widget.isLoggedInMandatory,
-    )) {
-      return;
-    }
-
-    if (!context.mounted) {
-      return;
-    }
-    final bool? confirmed = await showDialog<bool>(
-      context: context,
-      builder: (BuildContext context) {
-        return SmoothAlertDialog(
-          title: appLocalizations.confirm_button_label,
-          body: Text(
-            appLocalizations.are_you_sure,
-          ),
-          close: true,
-          positiveAction: SmoothActionButton(
-            text: appLocalizations.yes,
-            onPressed: () => Navigator.of(context).pop(true),
-          ),
-          negativeAction: SmoothActionButton(
-            text: appLocalizations.no,
-            onPressed: () => Navigator.of(context).pop(false),
-          ),
-        );
-      },
-    );
-    if (confirmed == true) {
-      if (context.mounted) {
-        final LocalDatabase localDatabase = context.read<LocalDatabase>();
-        await BackgroundTaskUnselect.addTask(
-          barcode,
-          imageField: widget.imageField,
-          context: context,
-          language: widget.language,
-        );
-        localDatabase.notifyListeners();
-        navigatorState.pop();
-      }
-    }
-  }
-
-  Future<File?> _openCropPage(
-    final NavigatorState navigatorState,
-    final File imageFile, {
-    final int? imageId,
-    final Rect? initialCropRect,
-    final CropRotation? initialRotation,
-  }) async =>
-      navigatorState.push<File>(
-        MaterialPageRoute<File>(
-          builder: (BuildContext context) => CropPage(
-            language: widget.language,
-            barcode: barcode,
-            imageField: _imageData.imageField,
-            inputFile: imageFile,
-            imageId: imageId,
-            initiallyDifferent: false,
-            initialCropRect: initialCropRect,
-            initialRotation: initialRotation,
-            isLoggedInMandatory: widget.isLoggedInMandatory,
-          ),
-          fullscreenDialog: true,
-        ),
-      );
-
-  Future<File?> _openEditCroppedImage(
-    final int imageId,
-    final ProductImage productImage,
-  ) async {
-    final NavigatorState navigatorState = Navigator.of(context);
-    final LocalDatabase localDatabase = context.read<LocalDatabase>();
-    final File? imageFile = await downloadImageUrl(
-      context,
-      ProductImage.raw(
-        imgid: imageId.toString(),
-        size: ImageSize.ORIGINAL,
-      ).getUrl(barcode),
-      DaoInt(localDatabase),
-    );
-    if (imageFile == null) {
-      return null;
-    }
-    return _openCropPage(
-      navigatorState,
-      imageFile,
-      imageId: imageId,
-      initialCropRect: _getCropRect(productImage),
-      initialRotation: CropRotationExtension.fromDegrees(
-        productImage.angle?.degree ?? 0,
-      ),
-    );
-  }
-
-  ProductImage? _getBestProductImage() {
-    if (upToDateProduct.images == null) {
-      return null;
-    }
-    for (final ProductImage productImage in upToDateProduct.images!) {
-      if (productImage.field != _imageData.imageField) {
-        continue;
-      }
-      if (productImage.language != widget.language) {
-        continue;
-      }
-      if (productImage.size == ImageSize.ORIGINAL) {
-        if (productImage.imgid != null) {
-          return productImage;
-        }
-      }
-    }
-    return null;
-  }
-
-  /// Returns a crop rect, to be compared with the full image dimensions.
-  Rect? _getCropRect(final ProductImage productImage) =>
-      productImage.x1 == null ||
-              productImage.y1 == null ||
-              productImage.x2 == null ||
-              productImage.y2 == null
-          ? null
-          : Rect.fromLTRB(
-              productImage.x1!.toDouble(),
-              productImage.y1!.toDouble(),
-              productImage.x2!.toDouble(),
-              productImage.y2!.toDouble(),
-            );
 }
